Split Labs screen into per-feature panels; one TTL parse path
LabsScreen (929 lines, 28 useState in four disjoint clusters) becomes
a thin composer owning only the shared form/plan/subscription/result:
LabInventoryPanel (list/extend/destroy), LabProvisionerPanels (profile,
target+TTL, plan review + permission check, deploy), FlowLogPackPanel
(pack assembly + install). New pure labs-state functions close the
seams the component was hand-rolling: labTtlFromForm (the deploy now
uses the SAME strict parse as the plan preview - the raw Number() path
is gone), parseExtendHours, packStorageAccountName (typed > deployed >
planned precedence, used by both the screen default and the panel).
